requirements.txt
autopep8==2.0.2
django-extensions==3.2.3
django-widget-tweaks==1.4.12
Django==4.2.3
djhtml==3.0.6
isort==5.12.0
python-decouple==3.8
dr-scaffold==2.1.2
django-ninja==0.22.2
psycopg2-binary==2.9.6
python-gitlab==3.15.0
rich==13.6.0
openpyxl==3.1.2
django-import-export==3.3.5
click==8.1.7
ruff==0.11.0
